1.1
1.1

+++ src/com/streamsdk/cache/MessagingAckDB.java
@@ -17,7 +17,7 @@ import android.util.Log;
 public class MessagingAckDB {
 
 	private static final String DATABASE_NAME = "mrdb";
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 2;
     private SQLiteDatabase db;
     private SQLiteStatement mInsertStmt;
 	private DatabaseHelper helper;
@@ -173,10 +173,16 @@ public class MessagingAckDB {
 			 db.execSQL("CREATE TABLE mrdb (id TEXT PRIMARY KEY, type TEXT, fromuser TEXT, touser TEXT, message TEXT,fileid TEXT, duration TEXT, tid TEXT, lat TEXT, lon TEXT, address TEXT)");
 		}
 
-		public void onUpgrade(SQLiteDatabase db, int arg1, int arg2) {
-			 db.execSQL("DROP TABLE IF EXISTS mrdb");
-	         onCreate(db);
-		}
+        public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        	if (oldVersion == 1 && newVersion == 2){
+        	  db.execSQL("ALTER TABLE mrdb ADD COLUMN lat TEXT");
+        	  db.execSQL("ALTER TABLE mrdb ADD COLUMN lon TEXT");
+        	  db.execSQL("ALTER TABLE mrdb ADD COLUMN address TEXT");
+        	}else{
+        	  db.execSQL("DROP TABLE IF EXISTS mrdb");
+        	  onCreate(db);
+        	}
+        }
     }
